requirements.txt
fastapi>=0.95.0,<0.100.0
uvicorn>=0.22.0,<0.24.0
pydantic>=1.10.0,<2.0.0
python-dotenv>=0.21.0
duckdb>=0.8.0
redis>=4.5.0
pandas>=1.5.0
polars>=0.18.0
openpyxl>=3.1.0
yfinance>=0.2.18
alpha_vantage>=2.3.1
pytest>=7.3.1
lxml>=4.9.0
html5lib>=1.1.0
beautifulsoup4>=4.11.0
